Updated to match current usage
Shortened the incomplete message because of how it's being used. Removed "Please edit this when editing the page" because nobody ever changes it, the template is simply removed altogether after a few days. The new version is clearer. Also, it works better with the refreshed {{incomplete}} and {{incomplete transcript}} templates, which have different text than before as they've been changed recently.